requirements.txt
torch>=1.1
torchvision
numpy
hydra-core>=1.0.3
tqdm
tensorboard>=1.14
ray
optuna
sklearn
